Update README.md
The files contain the data and calculations used in the paper, as follows.  CAfibroGH.Rmd is the main file for the paper's chromosome aberration (CA) sections; CAsynergyGH.R is an excerpt; HGsynergyMain.R is for the murine Harderian gland (HG) sections; Lea.R is for the supplementary information to the paper. These scripts can generate the data-based figures in the paper. The scripts, especially the commenting, will need to be updated, clarified and shortened according to reviewers' and editors' comments as the review process unfolds.